+++ music_collection/accounts/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
 
+from .models import Profile
 
+
+admin.site.register(Profile)
