+++ music_collection/accounts/models.py
@@ -1,78 +1,15 @@
+from django.contrib.auth.models import User
 from django.db import models
-from django.contrib.auth.models import (
-    BaseUserManager, AbstractBaseUser
-)
-
-class UserManager(BaseUserManager):
-    def create_user(self, email, password=None):
-        if not email:
-            raise ValueError('Users must have an email address')
-
-        user = self.model(
-            email=self.normalize_email(email),
-        )
-
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_staffuser(self, email, password):
-        user = self.create_user(
-            email,
-            password=password,
-        )
-        user.staff = True
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, email, password):
-        user = self.create_user(
-            email,
-            password=password,
-        )
-        user.staff = True
-        user.admin = True
-        user.save(using=self._db)
-        return user
 
 
-
-class User(AbstractBaseUser):
-    email = models.EmailField(
-        verbose_name='email address',
-        max_length=255,
-        unique=True,
-    )
-    is_active = models.BooleanField(default=True)
-    staff = models.BooleanField(default=False) # a admin user; non super-user
-    admin = models.BooleanField(default=False) # a superuser
-    objects = UserManager()
-
-
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = [] # Email & Password are required by default.
-
-    def get_full_name(self):
-        return self.email
-
-    def get_short_name(self):
-        return self.email
+class Profile(models.Model):
+    user = models.OneToOneField(User, null=True, on_delete=models.CASCADE)
+    about_me = models.TextField()
+    profile_pic = models.ImageField(upload_to="music_collection/my_media/profile_pic", null=True, blank=True)
+    facebook_url = models.CharField(max_length=255, null=True, blank=True, )
+    instagram_url = models.CharField(max_length=255, null=True, blank=True, )
+    linkedin_url = models.CharField(max_length=255, null=True, blank=True, )
 
     def __str__(self):
-        return self.email
+        return str(self.user)
 
-    def has_perm(self, perm, obj=None):
-        # Simplest possible answer: Yes, always
-        return True
-
-    def has_module_perms(self, app_label):
-        return True
-
-    @property
-    def is_staff(self):
-        return self.staff
-
-    @property
-    def is_admin(self):
-        return self.admin
