+++ music_collection/accounts/templates/registration/edit_profile.html
@@ -7,7 +7,7 @@
 <h1>Edit profile...</h1>
 <br/><br/>
 
-<img src={{user.profile.profile_pic}} width = 200, height=200, class='rounded-circle'>
+<img src="{{user.profile.profile_pic}}", width = 200, height=200, class='rounded-circle'>
 <br/><br/>
 
 <div>{{ user.profile.about_me }}</div>
