+++ music_collection/accounts/urls.py
@@ -1,13 +1,15 @@
-from django.urls import path
-from django.contrib.auth import views as auth_views
-
+from django.contrib import admin
+from django.urls import path, include
 from .views import UserRegisterView, UserEditView, PasswordsChangeView
+from django.contrib.auth import views as auth_views
 from . import views
 
+
 urlpatterns = [
-    path('register/', UserRegisterView.as_view(), name='register'),
-    path('edit_profile/', UserEditView.as_view(), name='edit_profile'),
-    path('password/', PasswordsChangeView.as_view(template_name='registration/change-password.html'),
-         name='change_password'),
-    path('password_success', views.password_success, name='password_success'),
+        path('register/', UserRegisterView.as_view(), name='register'),
+        path('edit_profile/', UserEditView.as_view(), name='edit_profile'),
+        path('password/', PasswordsChangeView.as_view(template_name='registration/change-password.html')),
+        path('password_success', views.password_success, name='password_success'),
+
+
 ]
