+++ music_collection/accounts/views.py
@@ -1,10 +1,10 @@
-from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import UserCreationForm, UserChangeForm, PasswordChangeForm
 from django.contrib.auth.views import PasswordChangeView
 from django.shortcuts import render
 from django.urls import reverse_lazy
 from django.views import generic
-from .forms import SignUpForm, RegisterForm, EditProfileView, PasswordChangingForm
+from .forms import SignUpForm, EditProfileForm, PasswordChangingForm
+
 
 class PasswordsChangeView(PasswordChangeView):
     form_class = PasswordChangingForm
@@ -14,13 +14,13 @@ def password_success(request):
     return render(request, 'registration/password_success.html', {})
 
 class UserRegisterView(generic.CreateView):
-    form_class = RegisterForm
+    form_class = SignUpForm
     template_name = 'registration/register.html'
     success_url = reverse_lazy('login')
 
 
 class UserEditView(generic.UpdateView):
-    form_class = EditProfileView
+    form_class = EditProfileForm
     template_name = 'registration/edit_profile.html'
     success_url = reverse_lazy('home')
