+++ music_collection/music/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import MusicAlbum, Category, Availability, Rating, Performer, Profile #profie should go to accounts app
+from .models import MusicAlbum, Category, Availability, Rating, Performer
 
 
 admin.site.register(MusicAlbum)
@@ -8,4 +8,4 @@ admin.site.register(Category)
 admin.site.register(Availability)
 admin.site.register(Rating)
 admin.site.register(Performer)
-admin.site.register(Profile)
+
